requirements.txt
Django==1.11
djangorestframework==3.6.2
mock=2.0.0
django-filter==1.0.2
